+++ sdk/storage/azure_storage_blob/src/models/mod.rs
@@ -3,6 +3,7 @@
 
 //! Model types for Azure Blob Storage.
 
+mod download_result;
 pub(crate) mod drains;
 pub mod error;
 mod extensions;
@@ -13,6 +14,8 @@ pub(crate) mod slice;
 mod upload_result;
 
 pub use crate::generated::models::*;
+pub use download_result::{BlobClientDownloadResult, BlobDownloadProperties};
+pub use method_options::BlobClientDownloadOptions;
 pub use method_options::BlockBlobClientUploadOptions;
 pub use method_options::BlockBlobClientUploadOptions as BlobClientUploadOptions;
 pub use upload_result::BlockBlobClientUploadResult;
@@ -118,13 +121,6 @@ pub mod blob_name {
         }
     }
 }
-
-// TODO: Need mechanism to have these models not have "Internal" suffix
-pub use crate::generated::models::{
-    BlobClientDownloadInternalOptions as BlobClientDownloadOptions,
-    BlobClientDownloadInternalResult as BlobClientDownloadResult,
-    BlobClientDownloadInternalResultHeaders as BlobClientDownloadResultHeaders,
-};
 
 #[cfg(test)]
 mod tests {