+++ sdk/storage/azure_storage_blob/tests/streaming.rs
@@ -7,11 +7,7 @@ use azure_core::{
     Bytes,
 };
 use azure_core_test::{recorded, stream::GeneratedStream, TestContext};
-use azure_storage_blob::{
-    format_page_range,
-    models::{BlobClientDownloadResultHeaders, BlockLookupList},
-    BlobClient,
-};
+use azure_storage_blob::{format_page_range, models::BlockLookupList, BlobClient};
 use azure_storage_blob_test::{get_blob_name, get_container_client, StorageAccount};
 use futures::TryStreamExt as _;
 use std::error::Error;
@@ -54,11 +50,12 @@ async fn stream_blob_upload(ctx: TestContext) -> Result<(), Box<dyn Error>> {
 
     // Assert
     let response = blob_client.download(None).await?;
-    let content_length = response.content_length()?;
-    let (status_code, _, response_body) = response.deconstruct();
-    assert!(status_code.is_success());
-    assert_eq!(data.len() as u64, content_length.unwrap());
-    assert_eq!(data.to_vec(), response_body.collect().await?.to_vec());
+    assert_eq!(
+        data.len() as u64,
+        response.properties.content_length.unwrap()
+    );
+    let body_data = response.body.collect().await?;
+    assert_eq!(data.to_vec(), body_data);
 
     container_client.delete(None).await?;
     Ok(())
@@ -108,14 +105,15 @@ async fn stream_stage_block(ctx: TestContext) -> Result<(), Box<dyn Error>> {
 
     // Assert
     let response = blob_client.download(None).await?;
-    let content_length = response.content_length()?;
-    let (status_code, _, response_body) = response.deconstruct();
-    assert!(status_code.is_success());
     let expected_len = block_1_data.len() + block_2_data.len();
-    assert_eq!(expected_len as u64, content_length.unwrap());
+    assert_eq!(
+        expected_len as u64,
+        response.properties.content_length.unwrap()
+    );
     let mut expected = block_1_data.to_vec();
     expected.extend_from_slice(block_2_data);
-    assert_eq!(expected, response_body.collect().await?.to_vec());
+    let body_data = response.body.collect().await?;
+    assert_eq!(expected, body_data);
 
     container_client.delete(None).await?;
     Ok(())
@@ -152,14 +150,15 @@ async fn stream_append_block(ctx: TestContext) -> Result<(), Box<dyn Error>> {
 
     // Assert
     let response = blob_client.download(None).await?;
-    let content_length = response.content_length()?;
-    let (status_code, _, response_body) = response.deconstruct();
-    assert!(status_code.is_success());
     let expected_len = block_1.len() + block_2.len();
-    assert_eq!(expected_len as u64, content_length.unwrap());
+    assert_eq!(
+        expected_len as u64,
+        response.properties.content_length.unwrap()
+    );
     let mut expected = block_1.to_vec();
     expected.extend_from_slice(block_2);
-    assert_eq!(expected, response_body.collect().await?.to_vec());
+    let body_data = response.body.collect().await?;
+    assert_eq!(expected, body_data);
 
     container_client.delete(None).await?;
     Ok(())
@@ -190,11 +189,9 @@ async fn stream_upload_pages(ctx: TestContext) -> Result<(), Box<dyn Error>> {
 
     // Assert
     let response = blob_client.download(None).await?;
-    let content_length = response.content_length()?;
-    let (status_code, _, response_body) = response.deconstruct();
-    assert!(status_code.is_success());
-    assert_eq!(512, content_length.unwrap());
-    assert_eq!(data, response_body.collect().await?.to_vec());
+    assert_eq!(512, response.properties.content_length.unwrap());
+    let body_data = response.body.collect().await?;
+    assert_eq!(data, body_data);
 
     container_client.delete(None).await?;
     Ok(())
@@ -224,7 +221,7 @@ async fn upload<const CONTENT_LENGTH: usize>(client: &BlobClient) -> azure_core:
 #[tracing::instrument(skip_all, fields(content_length), err)]
 async fn download(client: &BlobClient) -> azure_core::Result<u64> {
     let mut len = 0;
-    let mut response = client.download(None).await?.into_body();
+    let mut response = client.download(None).await?.body;
     while let Some(data) = response.try_next().await? {
         tracing::debug!("received {} bytes", data.len());